requirements.txt
Django==4.2.2
djangorestframework
markdown
django-filter
django-cors-headers
djangorestframework-simplejwt
stripe
psycopg
psycopg2
django-dotenv
python-dotenv